webkit-efl/LayoutTests/platform/efl-tizen/fast/images/imagemap-scroll-expected.txt

This tests to be sure that focusing an area element triggers scrolling and removing focus from it does not.

PASS: Document is starting scrolled to top.

FAIL: Document is scrolled to 0 after focusing area element.

PASS: Document is scrolled to top once again.

PASS: Document is still scrolled to top after removing focus from area element.

TEST COMPLETE
